+++ pkg/hmenum/rega_script.go
@@ -49,6 +49,7 @@ const (
 	RegaScriptGetInboxDevices       RegaScript = "get_inbox_devices"
 	RegaScriptTriggerFirmwareUpdate RegaScript = "trigger_firmware_update"
 	RegaScriptRebootCCU             RegaScript = "reboot_ccu"
+	RegaScriptSetCCUPosition        RegaScript = "set_ccu_position"
 	RegaScriptCreateBackupStart     RegaScript = "create_backup_start"
 	RegaScriptCreateBackupStatus    RegaScript = "create_backup_status"
 	RegaScriptUsageBySysvar         RegaScript = "usage_by_sysvar"
@@ -92,6 +93,7 @@ var AllRegaScripts = []RegaScript{
 	RegaScriptGetInboxDevices,
 	RegaScriptTriggerFirmwareUpdate,
 	RegaScriptRebootCCU,
+	RegaScriptSetCCUPosition,
 	RegaScriptCreateBackupStart,
 	RegaScriptCreateBackupStatus,
 	RegaScriptUsageBySysvar,